+++ rewrite-gradle/src/test/java/org/openrewrite/gradle/RemoveDependencyTest.java
@@ -26,7 +26,7 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.openrewrite.gradle.Assertions.buildGradle;
-import static org.openrewrite.gradle.Assertions.withToolingApi;
+import static org.openrewrite.gradle.toolingapi.Assertions.withToolingApi;
 
 class RemoveDependencyTest implements RewriteTest {
